\name{correction_factor}
\alias{correction_factor}
\docType{data}
\title{Gene average expression level and standard deviation in GTEx data
}
\description{Gene average expression level and standard deviation in GTEx data
}
\usage{data("correction_factor")}
\format{
  A data frame with 14725 observations on the following 2 variables.
  \describe{
    \item{\code{avg.all}}{a factor with gene average expression level }
    \item{\code{sd.all}}{a factor with gene standard deviation of expression level }
  }
}
\details{nothing
}
\source{nothing
}
\references{Pei G., Dai Y., Zhao Z., Jia P. (2019) Tissue-Specific Enrichment Analysis deTS: Tissue-Specific Enrichment Analysis to decode tissue specificity. Bioinformatics, In submission.
}
\examples{
data(correction_factor)
}
\keyword{datasets}
