\name{GWAS_gene}
\alias{GWAS_gene}
\docType{data}
\title{Gene symbol query data for single sample
}
\description{An example of input gene symbol query data for single sample tissue-specific enrichment analysis
}
\usage{data("GWAS_gene")}
\format{
  The format is:
  
	"A1BG" "A1BG-AS1" "A1CF" "A2M" "A2M-AS1" "A2ML1" "A2MP1" "A3GALT2" "A4GALT" "A4GNT" "AA06" "AAAS" "AACS" "AACSP1" "AADAC" ...
}
\details{nothing
}
\source{nothing
}
\references{Pei G., Dai Y., Zhao Z., Jia P. (2019) deTS: Tissue-Specific Enrichment Analysis to decode tissue specificity. Bioinformatics, In submission.
}
\examples{
data(GWAS_gene)
}
\keyword{datasets}
